+++ packages/sdk/src/common/chains.ts
@@ -111,6 +111,17 @@ export type SuiChain =
   | typeof SUI_LOCALNET_CHAIN
   | typeof SUI_MAINNET_CHAIN;
 
+export const isSuiChain = (chainId: unknown): chainId is SuiChain => {
+  return (
+    [
+      SUI_DEVNET_CHAIN,
+      SUI_TESTNET_CHAIN,
+      SUI_LOCALNET_CHAIN,
+      SUI_MAINNET_CHAIN,
+    ] as string[]
+  ).includes(chainId as string);
+};
+
 export const SOLANA_MAINNET_CHAIN = 'solana:mainnet-beta' as const;
 export const SOLANA_TESTNET_CHAIN = 'solana:testnet' as const;
 export const SOLANA_DEVNET_CHAIN = 'solana:devnet' as const;
@@ -119,6 +130,16 @@ export type SolanaChain =
   | typeof SOLANA_MAINNET_CHAIN
   | typeof SOLANA_TESTNET_CHAIN
   | typeof SOLANA_DEVNET_CHAIN;
+
+export const isSolanaChain = (chainId: unknown): chainId is SolanaChain => {
+  return (
+    [
+      SOLANA_MAINNET_CHAIN,
+      SOLANA_TESTNET_CHAIN,
+      SOLANA_DEVNET_CHAIN,
+    ] as string[]
+  ).includes(chainId as string);
+};
 
 export const ChainId = {
   ethereum: 1,